+++ CHANGELOG.rst
@@ -4,9 +4,10 @@ test_support Change Log
 UNRELEASED
 ----------
 
+  * ADDED:     Methods in Xsi class for getting the xsim tick frequency
   * CHANGED:   Pyxsim CMake build uses XCommon CMake
   * CHANGED:   The way time is incremented by time_step for better floating point precision
-  * ADDED:     Methods in Xsi class for getting the xsim tick frequency
+  * FIXED:     Resolved issues with stdout/stderr capture in Pyxsim
 
 2.0.0
 -----